Value of bit

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -2,10 +2,11 @@
 
 /**
  * get_bit - returs the value of a bit at agiven index
+ * @n: bit to be evaluated
  * @index: the index starting from 0 of the bit you want to grt
  * Return: the value of the bit at index
  */
-int get_bit(unsigned long int n, unsigned long int index)
+int get_bit(unsigned long int n, unsigned int index)
 {
 	if (index > 63)
 		return (-1);
